More consistently use the words "seqNum" (which is log terminology), "index" (which is replication terminology).
The OLog interface, OLogEntry objects, all code behind the OLog interface, and all code testing that code, use the word "seqNum". This is to avoid confusion with the other meaning of the word "index" meaning "meta-information to make lookups faster".

Whereas the Replicator and ReplicatorLog interfaces, IndexCommitNotices, LogEntry messages, the ReplicatorInstance, and all code testing it, refer to "index", in order to agree with the language used in discussing the consensus algorithm.

Mooring, the boundary between replication and log, translates one word to the other.